define custom assert function, add extra argument to scripts

+++ headers/util.hpp
@@ -2,7 +2,7 @@
 #include <vector>
 #include <cstdlib>
 
-double assert_stretch_bound(std::vector<std::vector<double>> dist_original, std::vector<std::vector<double>> dist_stretch, int stretch_factor);
+double assert_stretch_bound(std::vector<std::vector<double>> dist_original, std::vector<std::vector<double>> dist_stretch, double stretch_factor);
 void print_matrix(std::vector<std::vector<double>> matrix);
 
 inline double random_number() {
